add docker support for v2-push
also adds some basic validation for the flags themselves

[Finishes #143337915]

+++ actor/pushaction/apply.go
@@ -71,30 +71,34 @@ func (actor Actor) Apply(config ApplicationConfig, progressBar ProgressBar) (<-c
 			eventStream <- BoundRoutes
 		}
 
-		archivePath, err := actor.CreateArchive(config)
-		if err != nil {
-			errorStream <- err
-			return
-		}
-		eventStream <- CreatingArchive
-		defer os.Remove(archivePath)
-
-		for count := 0; count < PushRetries; count++ {
-			warnings, err = actor.UploadPackage(config, archivePath, progressBar, eventStream)
-			warningsStream <- warnings
-			if _, ok := err.(ccerror.PipeSeekError); !ok {
-				break
-			}
-			eventStream <- RetryUpload
-		}
-
-		if err != nil {
-			if _, ok := err.(ccerror.PipeSeekError); ok {
-				errorStream <- UploadFailedError{}
+		if config.DesiredApplication.DockerImage == "" {
+			archivePath, err := actor.CreateArchive(config)
+			if err != nil {
+				errorStream <- err
 				return
 			}
-			errorStream <- err
-			return
+			eventStream <- CreatingArchive
+			defer os.Remove(archivePath)
+
+			for count := 0; count < PushRetries; count++ {
+				warnings, err = actor.UploadPackage(config, archivePath, progressBar, eventStream)
+				warningsStream <- warnings
+				if _, ok := err.(ccerror.PipeSeekError); !ok {
+					break
+				}
+				eventStream <- RetryUpload
+			}
+
+			if err != nil {
+				if _, ok := err.(ccerror.PipeSeekError); ok {
+					errorStream <- UploadFailedError{}
+					return
+				}
+				errorStream <- err
+				return
+			}
+		} else {
+			log.WithField("docker_image", config.DesiredApplication.DockerImage).Debug("skipping file upload")
 		}
 
 		configStream <- config
